requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
requests==2.31.0
flask-socketio==5.3.6
python-socketio==5.10.0
websocket-client==1.7.0